Added camera movements and clipping (large commit, read desc)
Camera can now move up, down, left, right, forwards, backwards, and rotate left and right.

Note: the left and right movement is fixed to the x-axis. So if camera rotates right then move's forwards, instead of moving to it's own left or right, it will still move along the x-axis. The same thing would presumably happen to up and down movement if rotating the camera up and down was implemented.

Meshes are still rendered properly when parts of the mesh (or the whole thing) are off the screen due to clipping.

Currently in the middle of refactoring rendering code, so the Renderer header and source files are unfinished.