Fixed issue where object was not rendered to screen if no rotation was applied on any of it's axes
This was caused by passing in a 0 vector to the MatrixMakeRotation() function. Since it first normalizes the vector, trying to normalize a zero vector results in a divide by 0, which is undefined behavior.
To solve this, I implemented a simple check in the function. If the passed in axis is a zero vector, then return the identity matrix.