Fixed camera movement
The issue was that I didn't pass the reference to the cameraComponent. Simple fix.

However now that it does work, the game seems to be slower than before, and it gets alot more slow  when another entity is added. I assume it's mostly from the rendering system, so some optimizing will have to be done.